Adding support for OpenSSL engines, allowing keys to be in an HSM
The previous code required keys to be stored on disk.  This patch allows
using keys that are stored in an HSM.  For example, CloudHSM has a fake
key that only contains the key identifier, which is what is stored on
disk and passed into fsverity.  The engine knows how to use the fake key
to request the HSM use the real key to do the signing operation.  Thus
the user never has access to the key material, but merely the ability to
use it.